Show Subcloud Deploy Configuration Status
This commit extends the
"dcmanager subcloud show <subcloud> --detail" command to display
the deploy configuration status of the subcloud from the
systemcontroller.

The presence of a 260.002 alarm related to a 'host' resource on an
online subcloud indicates an outdated deploy configuration.
In such cases, the field will display "Deployment: configurations
out-of-date". Otherwise, the field  will show:
"Deployment: configurations up-to-date". The field will not be shown
for offline subclouds.

Test Plan:
PASS: Successful bring up of a fresh DC system with these changes.
PASS: Run 'dcmanager subcloud show' without the '--detail' argument.
Verify that the new field is not being shown.
PASS: Run 'dcmanager subcloud show --detail' for an online subcloud
containing a 260.002 alarm related to a host resource. Verify that
the output contains 'Deployment: configurations out-of-date'.
PASS: Run 'dcmanager subcloud show --detail' for an online subcloud
containing a 260.002 alarm related to a resource other than host.
Verify that the output contains 'Deployment: configurations
up-to-date'.
PASS: Run 'dcmanager subcloud show --detail' for an online subcloud
without a 260.002 alarm. Verify that the output contains 'Deployment:
configurations up-to-date'.
PASS: Run 'dcmanager subcloud show --detail' for an offline subcloud.
Verify that the new field is not being shown.

Story: 2010719
Task: 48025

Change-Id: I223919ded1d6d5dfcc81c2060c5e6491ed8a73d9
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>