Extend subcloud bootstrap log rotation size
This commit extends the file size for ansible bootstrap log rotation.
The size is extended so that the file is not rotated during the
subcloud bootstrap operation.

The new size (450K) to keep the file as small as possible (the file is
uncompressed, with at least one per deployed subcloud), with a margin
for expanded log size in the future.

Test Plan:
PASS: Verify the ansible log is still rotated but not during the
first bootstrap operation, and that the other rotation parameters are
not affected.

Story: 2010611
Task: 48319

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I0fa839dedde4eedecb922e7aa1cfab9b56bc0fd6