Rename config_sync_status field
This commits updates the 'config_sync_status' field name to
'deploy_config_sync_status' in order to distiguish from the
config-out-of-date information related to the 250.001 alarm.

Test Plan:
PASS: Build distributedcloud package. Install it on a DC system
alongside the client changes. Verify that the client receives
the updated field name and that the name printed on the output
command is the expected.
PASS: Verify the correct functionality of the 'dcmanager
subcloud show --detail' command.

Story: 2010719
Task: 48089

Change-Id: I8e573947dfa61700f15bba4313939bb7a7959d53
Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>