Set the correct deploy_status value in subcloud restore response
In this commit, the deploy_status value is set to pre-restore in
the subcloud-backup restore response as opposed to the last
deploy_status value saved in the database.

Test Case:
PASS: run subcloud-backup restore command for a group of subclouds
and verify that deploy_status value is pre-restore in
the command output.
PASS: run subcloud-backup for a group of subclouds that have
deploy_status restore failed and verify that deploy_status value
is pre-restore in the command output.
PASS: run subcloud-backup restore command for a single subcloud
and verify that deploy_status value is pre-restore in
the command output.

Closes-Bug: 2023216

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: Ic1d113dcb855e80e8e8473fb1f6deeca079bef25