Move DC ansible override files to dc-vault
This commit moves the ansible override files from /var/opt/dc/ansible
to /opt/dc-vault/ansible, which is replicated across both controllers.

This is done to address issues that arise when a subcloud is first
created on one controller, followed by a swact, and the user attempts
to perform another operation (such as subcloud reconfig). This
situation leads to operation failures because the necessary ansible
override files, such as the subcloud inventory file, are not available
on the current active controller.

Test Plan:
1. PASS - Verify that /opt/dc-vault/ansible directory gets created
          automatically with the correct mode (600) and owned by root;
2. PASS - Create a subcloud on controller-0 with 'dcmanager subcloud
          deploy create', passing the bootstrap-values, deploy-config
          and install-values parameters, swact to controller-1 and then
          run the deploy install, bootstrap and config without passing
          the parameters again. Verify that the subcloud is deployed
          correctly and that the operations are reading the files
          stored in dc-vault/ansible;
3. PASS - Deploy a subcloud with 'dcmanager subcloud add', passing the
          bootstrap-values, deploy-config and install-values,
          parameters, swact to controller-1 and run 'dcmanager subcloud
          reconfig' and verify that the operation completes
          successfully (the operation should use the subcloud inventory
          file stored inside dc-vault);

Story: 2010756
Task: 48243

Change-Id: I6ac25bcd07f19a14f1bb5523d913a9e51b708d35
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>