Set management_start_address on subcloud update
This commit fixes a recent issue which caused the service
endpoints on the system controller side to not be updated
properly when updating the networking parameters of a
subcloud.

As part of the refactoring of network configuration for
subcloud re-install, the management_start_address value
(or admin_start_address) in the payload is now used for
the determination of the services endpoint.

On subcloud update with network parameters, the equivalent
value in the payload is management_start_ip

Test plan:

- Update a subcloud to use new admin network parameters.
- Update a subcloud to fallback from an admin network
  to the existing management network.

Story: 2010319
Task: 48437

Change-Id: Ife5ec61c9fb4f8d94b4e4d05f3d41f72e8952c8e
Signed-off-by: Steven Webster <steven.webster@windriver.com>