Add a switch to dcmanager orchestrator
Add a switch to dcmanager orchestrator to allow the orchestrator
to use the new usm api. Use a service flag to define whether patching/upgrade
or usm api will be used by the orchestrator. Once the usm api is complete,
this switch will be removed.

Story: 2010676
Task: 48530
Depends-On: https://review.opendev.org/c/starlingx/distcloud/+/884442

Test Case:
PASS: turn the usm switch on and verify that the USM API is used.
PASS: turn the usm switch off and verify that the patch API is
used and the regular states are used.

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I3045f62f1541bacd2dd65392692432f4b0519dd1
Signed-off-by: Hugo Brito <hugo.brito@windriver.com>