Fix subcloud availability status during subcloud install
Currently the subcloud availability status shows as
online while the subcloud is offline e.g. subcloud
reinstalled over system restore or upgrade.

This commit allows audit when the subcloud is being
installed if it's not offline to begin with.
The deploy_status thus can be updated accordingly
upon state change over the course of installation.

Test Plan:

1. PASS - Deploy a subcloud, create a subcloud backup
and then run dcmanager subcloud-backup restore with
--with-install option. Verify that the subcloud
availability state is set to 'offline' when the
deploy-state is set as 'installing'.

2. PASS - Re-run previous test modifying the playbook
to induce an error before creating the RVMC pod.
Verify that the subcloud remains online.

3. PASS - Re-run subcloud deploy, backup and restore.
Verify that the subcloud gets audited when it's
online and installing. Once it becomes offline during
the install phase, verify that the audit is skipped.

Closes-bug: 2031118

Change-Id: Ic206f23b2755454d163b78502825a77d03b3d86e
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>