Add a switch to patching proxy
Add a switch to patching to allow the proxy to handle patch api
and usm requests. Once the usm is fully integrated,
this parameter will be removed(Task 48986).

Depends-on: https://review.opendev.org/c/starlingx/stx-puppet/+/891343
Story: 2010676
Task: 48615

Test Case:
PASS: Turn the switch on and upload a patch using the software
upload command and verify that the patch is uploaded to dc-vault
PASS: Turn the switch on and upload a patch using the software
upload-dir command and verify that the patch is uploaded to dc-vault
PASS: Turn the switch off and upload a patch using the sw-patch
upload command and verify that the patch is uploaded to dc-vault
PASS: Turn the switch off and upload a patch using the sw-patch
upload-dir command and verify that the patch is uploaded to dc-vault
PASS: Turn the switch off and delete a patch using the sw-patch
delete command and verify that the patch is removed from dc-vault
PASS: Turn the switch on and delete a patch using the software
delete command and verify that the patch is removed from dc-vault
PASS: turn the switch on and commit a patch and verify that
the usm status is set to unknown.
Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I202df4beb4979ddcbbf5a926b9b53e265aa77333