Properly log subcloud install cleanup script failure
This commit ensures we properly log any script failures from the
gen-bootloader-iso*.sh scripts when invoked for cleanup (--delete).
The subprocess output is captured via subprocess.run, and is fully
logged on failure.

Note: in some cases, the gen-bootloader-iso*.sh script doesn't catch all
non-zero exit codes. We may want to address this at some point. For now,
we just make sure to print out the log in the cleanup operation so that
any failures are visible in the log file.

Test Case:
PASS: Success path: install subcloud, test that cleanup operation is
properly executed, and only the invocation is logged
PASS: Failure path: install subcloud, test that cleanup operation is
executed with failure inserted the script --delete. Verify that the full
log output is available in dcmanager.log

Closes-Bug: 2031446

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I303e7bd817dc84954fa5880c05eaeb75c2bfcf9b