Skip load-import if usm is enabled
Update dcorch platform proxy to skip code related to system
load-import when usm switch is enabled. Once the usm is fully integrated, the load-import code will be removed.

Depends-on: https://review.opendev.org/c/starlingx/distcloud/+/891345
Story: 2010676
Task: 48646

Test Case:
PASS: turn the switch off and run system load-import with
region name systemcontroller and verify that the load is uploaded to
dc-vault.
PASS: turn the switch on and run system load-import with
region name systemcontroller and verify that the load is not
uploaded to dc-vault.
PASS: turn the switch off and run system load-delete with
region name systemcontroller and verify that the load is removed from dc-vault.
PASS: turn the switch on and run system load-delete with
region name systemcontroller and verify that the load is not removed from dc-vault.

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I557f230b281eac10e5c5f20c35d437166ed4818f