Populate passwords during subcloud rehome
When running subcloud add with the --migrate option the admin_password
and ansible_become_pass where not being added to the ansible overrides
file, causing the rehoming playbook to fail when ansible tries to
connect to the subcloud. This bug was introduced in commit:
https://review.opendev.org/c/starlingx/distcloud/+/888274

This commit correctly populates the payload with the password fields
so they are written to the ansible overrides file during the deploy
create function.

Test plan:
1. PASS: Rehome a subcloud with this fix applied, making sure the
         bootstrap_values.yml doesn't already contain the password
         fields. Verify that the rehoming completes successfully
         and that the resulting ansible overrides file contains the
         required fields;
2. PASS: Add a subcloud with 'dcmanager subcloud add' and verify
         that the operation still completes successfully;
3. PASS: Run 'dcmanager deploy create' and verify that it completes
         successfully (the passwords should not be included in the
         resulting ansible overrides file).

Story: 2010756
Task: 48336

Change-Id: I26dede8499768c54cb289a065ac07ef751964b43
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>