Create software upload state
Create software upload state for the upgrade orchestrator, the state
will upload the releases from the systemcontroller that are either
deployed or committed to the subcloud if the releases are not present.

Story: 2010676
Task: 48711

Test Case (Patching only):
PASS: Run the upgrade orchestration and verify that the orchestator
upload a release to the subcloud that was deployed on the
systemcontroller.
PASS: Run the upgrade orchestration and verify that the orchestrator
failed to upload a release to the subcloud because the release wasn't
in dc-vault.

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: Ieaff09ce6a2542cc14f533bd274a0c19298969fe
(cherry picked from commit 3aa2a7f91d764966fd0f253af28a0a67ac3aa9f3)