Add the --upload-only option to dcmanager upgrade_strategy
This commit adds a new option --upload-only to dcmanager
upgrade-strategy create. If this option is used, the releases
are uploaded to the subclouds and the strategy state is set to complete.

Depends-on: https://review.opendev.org/c/starlingx/distcloud/+/893425

Test Plan:
1. PASS - Create and apply a patch strategy using the --upload-only
option and verify that the strategy completes after uploading
the patches to the subclouds.
2. PASS - Create and apply a patch strategy without using the
--upload-only option and verify that the strategy continues
after uploading the patches to the subclouds.

Story: 2010676
Task: 48747

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I3941151e71e356e5e91dbbf931e2302116c40277
(cherry picked from commit f3817f6c4484f72f11ce731fef08a014b341263e)