Update dcmanager audit to use usm API
In this commit, a new subcloud_status called software was created, it's
used when the usm switch is enabled to audit patches and loads using
the usm API. When the usm switch is disabled, the subcloud_status patch
and load are audited instead, using patch/sysinv API. The validations
to create an upgrade orchestration strategy were changed when the usm
switch is enabled, the subcloud_status software is checked instead of
the load status.

Test Plan:
PASS: Turn the usm switch on and add a subcloud and verify that
the subcloud has the subcloud_status software.
PASS: Turn the usm switch off and verify that the patch/load status
are audited.
PASS: Turn the usm switch on and verify that the software status
is audited.
PASS: Turn the usm switch on, upload a patch to the subcloud and verify
that the software status goes to out-of-sync.
PASS: Turn the usm switch on, apply a patch to the system controller and verify
that the software status goes to out-of-sync.
PASS: Turn the usm switch on, with the software status out-of-sync,
create an upgrade-strategy and verify that is created successfully.
PASS: Turn the usm switch on, restart audit and verify that the usm endpoint
was added to a existing subcloud.

Story: 2010676
Task: 48784

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: If2d14c15a6ff4e38b004b24700d3443a2e86d2c2