Fix to prevent mishandling of OS_REGION_NAME
With the Subcloud Rename Feature, a mechanism was introduced in
dcmanager, when the deployment file is uploaded, to modify the
OS_REGION_NAME prior to send it to the subcloud, required for
the installation of the Deployment.

Doing this manipulation of the deployment file on the dcmanager
site generates confusion for the handling of the required
OS_REGION_NAME parameter.

This commit revers the change to handle the deployment file from
the dcmanager, passing the responsability to the Deployment.

Closes-Bug: 2036632

Test Plan:

PASS: Add a new subcloud and validate the content of
deployment_config.yaml inside of subcloud. The value of
OS_REGION_NAME parameter should be equal to the same parameter
taken from environment variables.

Change-Id: Ie50459930e968ab5b5c0940e71110b636df297a6
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>