Create finish software strategy
Create finish software strategy for the upgrade orchestrator,
this state will commit the patches from the subcloud that are committed
in systemcontroller and it will delete releases from the subcloud that
are available or unavailable and not present in the systemcontroller.

Story: 2010676
Task: 48825

Test Case (Patching only):
PASS: Apply and commit a patch to the systemcontroller and apply the
same patch to the subcloud and run the upgrade orchestration and
verify that the patch from the subcloud is committed.
PASS: Upload a patch to the subcloud and run the upgrade orchestration
and verify that the patch is deleted from the subcloud.
PASS: Upload a patch to subcloud, protect the patch metadata file
with chattr +i, create and apply a upgrade strategy and verify
that the orchestrator failed to delete that patch and the correct
error message is presented.

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I2dae16800235891fc863151b307313732e8d49cd