Enhance ssh connection params to avoid key verification
As part of Subcloud Name Reconfiguration feature a mechanism was
instruduced to retrieve the current region name from a subcloud
that is to be rehomed to another SystemController.

This mechanism is based on the implementation of ssh, but it did
not have the necessary controls for the connection to the subcloud
to be efficient, which generated a problem in recovering the value
of the region causing a ssh key verification issue, necessary to
start the rehoming process.

This commit adds an additional parameters to the ssh command to
prevent key verification when te System Controller tries to connect
to a subcloud that it is not aware of.

Closes-Bug: 2037095

Test Plan:

PASS: Rehome a new subcloud (unknown by System Controller) and
verify that the rehoming process finish successfully. This would
be a subcloud whose key does not exist within the known_hosts file.

PASS: Delete an existing subcloud and rehome it on the same System
Controller. It is assumed that the key should exist on system
where a subcloud was initially configured, so the rehoming can be
verified.

Change-Id: I5a89c7a143aa4d8f11a348f58b1e4eb47b92fdde
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>