Update prestage orchestration to use usm
In this commit, the prestage orchestration was updated to use
usm audit status to check if the prestage-strategy can be created.
The request to sysinv to get the load was replaced by the request to
usm when usm switch is on.

Test Plan:
PASS: turn the usm switch on, create and apply a prestage-strategy and
verify that the strategy was completed successfully.
PASS: turn the usm switch off, create and apply a prestage-strategy and
verify that the strategy was completed successfully.

Story: 2010676
Task: 48852

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: I5882fa099ada5e940d61212ebd51f26baa5bda83