Missing abort operation when subcloud is renamed in managed mode
As part of Subcloud Name Reconfiguration feature a validation was added
to check the management state in order to rename de subcloud.

The subcloud has to be in unmanaged and with a valid deploy state to be
renamed, otherwise the dcmanager will return the error message, but this
is not working as expected, the subcloud is renamed while in managed
mode.

This commit add the abort operation to prevent the subcloud from being
renamed.

Closes-Bug: 2037684

Test Plan:
PASS: With the subcloud being unmanaged and deploy complete, rename it
and verify that the subcloud can be renamed successfully.

PASS: With the subcloud being managed and deploy complete, rename it
and verify that the subcloud cannot be renamed.

PASS: With the subcloud being unmanaged and deploy not complete, rename
it and verify that the subcloud cannot be renamed.

PASS: With the subcloud being managed and deploy not complete, rename it
and verify that the subcloud cannot be renamed.

Change-Id: If1a4596a3bb39d6ef57ac05639f3a6c57d7dea79
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>