Load-import failed due to lack of space in /scratch
During a load-import operation, the iso file is copied to /scratch,
a temporary folder (https://docs.starlingx.io/planning/kubernetes/storage-planning-storage-on-worker-hosts.html#root-filesystem-storage).
Before the operation, it is checked if this folder has a free space of
at least 3 times the ISO file, if not the load-import is aborted.

This change enhances the message displayed to the user in order to
give him some context and two solutions:
  - Delete some unused files from /scratch
  - Increase /scratch size

Test Plan:
  PASS: After adding more free space to /scratch, the load-import
operation must pass without error

Related to: https://review.opendev.org/c/starlingx/config/+/899865
Partial-bug: 2031941

Change-Id: Iff99e8793ab6fa5c812648bba4d93e8eab22a6b5
Signed-off-by: Lindley Werner <Lindley.Vieira@windriver.com>