Improve subcloud get operation to avoid subcloud lookup by region name
As part of Subcloud Name Reconfiguration feature, a mechanism was added
to perform the subcloud lookup through the name or region, to obtain
the subcloud information of those services that required it.

A scenario arises when an existing subcloud is renamed, and keeps the
old name as the region name. When invoking to dcmanager subcloud show
command passing the previous name, the command returns the information.
This is because there is an OR condition to do the lookup.

In order to avoid this behavior and return the subcloud information
only when the current name is supplied, an improvement was added to
distinguish the origin of the request and not affect CRUD operations.

Story: 2010788
Task: 48921

Test Plan:
PASS: Having an existing subcloud with name subcloud1 and region name
      subcloud1. Rename the subcloud to new name running dcmanager
      subcloud update [current subcloud] --name [new name]. Then
      run dcmanager subcloud show [new name] to show subcloud
      information.
PASS: Following the previous test to see the dcmanager output when
      region name is provided, run dcmanager subcloud show old_name,
      the message should be displayed indicating that the subcloud
      does not exist.

Change-Id: I37f847e7dd79ba53b90f952ee1848eeef69bb464
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>