Support rehoming a CentOS subcloud to a Debian system controller
This commit adds the following support for the ansible subcloud
rehoming playbook from the previous CentOS release, to enable rehoming
a CentOS subcloud to a Debian system controller:

1. Exports the smapi password to the subcloud's overrides file
2. Adds the absolute path of the validate_keystone_password.sh on the
subcloud to support the playbook to invoke this script in the
subcloud.

Test plan:
1. PASS - Rehome an AIOSX subcloud installed with the same software
version with the system controllers.
2. PASS - Rehome an AIOSX subcloud installed with 21.12 version to
a system controller installed with the master, specifying the
--release argument with the value of 21.12.

Closes-bug: 2039507
Signed-off-by: Yuxing Jiang <Yuxing.Jiang@windriver.com>
Change-Id: I24367e88acf6f785827b27daa9d390b188ae423e