Send system-local-ca secret ca.crt (DX sc upgrade)
The sysinv API call for certificate installation with type
openldap_ca will extract ca data included in the certificate bundle
and include it in the 'system-local-ca' ca.crt field.

Modified dcmanager to perform the call using this structure, passing
a bundle with TLS cert + CA cert + TLS key from the 'system-local-ca'
in the SystemController.

This code is called during DX subcloud upgrade and is used to keep
the current 'system-local-ca' on the subcloud consistent with the
one in the SystemController.

Test plan:
PASS: In a DC w/ DX subcloud in stx 9:
      - Perform cert-manager migration.
      - Upgrade the SystemController.
      - Verify system-local-ca secret content in the
        SystemController and the subcloud.
      - Start orchestrated upgraded for th DX subcloud.
      - Verify dcmanager/state.log. After the step
        "Stage: 2, State: transferring CA certificate"
        verify the system-local-ca secret content in the subcloud.
        The secret should have been replaced to match the one in
        the SystemController.
      - While c1 is upgrading, verify OpenLDAP by creating user
        in SystemController with 'ldapusersetup' and log into it
        in the subcloud.

Story: 2009811
Task: 49044

Change-Id: I42e6308f066126f903738f4e3c319c6027c8cb0b
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>