Support k8s multi-version upgrade in dc orchestrator
In this commit, the kube orchestrator was modified to select
the latest available version. When "--to-version" is provided, return
the desired version if the version is available, otherwise
return the latest available version. The unit tests will be added
in a subsequent commit.

Test Plan:
1. PASS - Create and apply kube orchestration without --to-version
and verify that kube was upgraded to the system controller version.
2. PASS - Create and apply kube orchestration with --to-version
selecting the latest version available and verify that kube
was upgraded to the correct version.
3. PASS - Create and apply kube orchestration with --to-version
selecting the first version available and verify that the kube
was upgraded to the correct version.

Story: 900672
Task: 49089

Co-authored-by: Chris Friesen <chris.friesen@windriver.com>
Co-authored-by: Christopher de Oliveira Souza <Christopher.DeOliveiraSouza@windriver.com>

Signed-off-by: Christopher Souza <Christopher.DeOliveiraSouza@windriver.com>
Change-Id: Ib7d7b978d4e4bf9c51262c15d5f2f4dbf4c37773